store cooking progress for guest (#27)
* handle show recipe detail page not auth

* store cooking progress for guest

+++ app/Http/Controllers/ProgressController.php
@@ -7,13 +7,26 @@ use App\Models\StepProgress;
 use App\Models\UserToolProgress;
 use App\Models\UserIngredientProgress;
 use App\Models\Recipe;
+use Session;
+use Auth;
 
 class ProgressController extends Controller
 {
     public function resetCookingProgress(Request $req){
-        UserToolProgress::where('user_id', auth()->id())->where('recipe_id', $req->recipe_id)->delete();
-        UserIngredientProgress::where('user_id', auth()->id())->where('recipe_id', $req->recipe_id)->delete();
-        StepProgress::where('user_id', auth()->id())->where('recipe_id', $req->recipe_id)->delete();
+        if(Auth::user()){
+            UserToolProgress::where('user_id', auth()->id())->where('recipe_id', $req->recipe_id)->delete();
+            UserIngredientProgress::where('user_id', auth()->id())->where('recipe_id', $req->recipe_id)->delete();
+            StepProgress::where('user_id', auth()->id())->where('recipe_id', $req->recipe_id)->delete();
+        }else{
+            Session::forget('recipe_'.$req->recipe_id);
+        }
+            
+
         return response()->json(['message' => 'ok done reset cooking progress']);
+    }
+
+    public function guest_getCookingProgress(Request $req){
+        $recipe_id = $req->recipe_id;
+        return Session::get('recipe_'.$recipe_id);
     }
 }
